Refactor publish vue
